Moved DataPrepper to core package
