+++ data-prepper-core/src/main/java/com/amazon/dataprepper/DataPrepperExecute.java
@@ -5,6 +5,7 @@
 
 package com.amazon.dataprepper;
 
+import com.amazon.dataprepper.core.DataPrepper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
